CustomerUI/sematic_azure_ai_agents copy 2/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
azure-identity==1.15.0
semantic-kernel==1.0.0
azure-ai-projects==1.0.0b3
aiofiles==23.2.1
python-multipart==0.0.6
